+++ packages/contracts/contracts/voting/token/TokenVotingSetup.sol
@@ -33,8 +33,8 @@ contract TokenVotingSetup is PluginSetup {
     /// @notice The address of the `TokenVoting` base contract.
     TokenVoting private immutable tokenVotingBase;
 
-    /// @notice The address zero to be used as oracle address for permissions.
-    address private constant NO_ORACLE = address(0);
+    /// @notice The address zero to be used as condition address for permissions.
+    address private constant NO_CONDITION = address(0);
 
     /// @notice The address of the `GovernanceERC20` base contract.
     address public immutable governanceERC20Base;
@@ -197,7 +197,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
@@ -205,7 +205,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
         );
 
@@ -214,7 +214,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             _dao,
             plugin,
-            NO_ORACLE,
+            NO_CONDITION,
             DAO(payable(_dao)).EXECUTE_PERMISSION_ID()
         );
 
@@ -225,7 +225,7 @@ contract TokenVotingSetup is PluginSetup {
                 PermissionLib.Operation.Grant,
                 token,
                 _dao,
-                NO_ORACLE,
+                NO_CONDITION,
                 tokenMintPermission
             );
         }
@@ -268,7 +268,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
@@ -276,7 +276,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
-            NO_ORACLE,
+            NO_CONDITION,
             tokenVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
         );
 
@@ -284,7 +284,7 @@ contract TokenVotingSetup is PluginSetup {
             PermissionLib.Operation.Revoke,
             _dao,
             _plugin,
-            NO_ORACLE,
+            NO_CONDITION,
             DAO(payable(_dao)).EXECUTE_PERMISSION_ID()
         );
 
@@ -298,7 +298,7 @@ contract TokenVotingSetup is PluginSetup {
                 PermissionLib.Operation.Revoke,
                 token,
                 _dao,
-                NO_ORACLE,
+                NO_CONDITION,
                 GovernanceERC20(token).MINT_PERMISSION_ID()
             );
         }